+++ toolkit/views/encounter_generator/encounter_generator_view.py
