+++ toolkit/views/loot_generator/cache_loot.py
@@ -2,7 +2,7 @@ from typing import Optional
 
 from django.contrib.auth.models import User
 
-from toolkit.cache import get_cache
+from toolkit.utilities.cache import get_cache
 from toolkit.models import Armor, GeneratedLoot, GenericItem, MagicItem, Weapon
