+++ toolkit/views/loot_generator/loot_generator_backend.py
